% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{AllSpatialFeaturePlot}
\alias{AllSpatialFeaturePlot}
\title{AllSpatialFeaturePlot}
\usage{
AllSpatialFeaturePlot(
  object,
  normalizeMethod = c("SCT", "log"),
  integreation.method = c("wnn", "mcia", "intnmf", "tica"),
  features = NULL,
  cluster.frame = TRUE,
  cluster.highlight = NULL,
  col = NULL,
  ...
)
}
\description{
AllSpatialFeaturePlot
}
