% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoBayesSpace.R
\name{BayesSpaceChainMatrixSimilarity}
\alias{BayesSpaceChainMatrixSimilarity}
\title{BayesSpaceChainMatrixSimilarity}
\usage{
BayesSpaceChainMatrixSimilarity(bayesSpaceResultChainMatrix)
}
\arguments{
\item{bayesSpaceResultChainMatrix}{MCMC chain saved in object}
}
\value{
BayesSpace similarity matrix
}
\description{
BayesSpaceChainMatrixSimilarity
}
