% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityIntegration.R
\name{DistToNeighborObj}
\alias{DistToNeighborObj}
\title{DistToNeighborObj}
\usage{
DistToNeighborObj(distMatrix, k.param = 20)
}
\arguments{
\item{distMatrix}{distance matrix}

\item{k.param}{number of k}
}
\value{
a Neighbor object
}
\description{
DistToNeighborObj
}
