% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBayesSpace.R
\name{doBayesSpace}
\alias{doBayesSpace}
\title{doBayesSpace}
\usage{
doBayesSpace(
  dataObj,
  platform = c("ST", "Visium"),
  assay = c("SCT", "Spatial", "ImageFeature"),
  name = paste0(assay, "BayesSpace_Cluster", nCluster),
  graphs.name = paste0(name, c("_nn", "_snn")),
  graphNameSimilarity = paste0(name, c("_Similarity")),
  graphNameDist = paste0(name, c("_Dist")),
  graphNameChain = paste0(name, c("_ChainMatrix")),
  nn.name = paste0(name, c("_Neighbor")),
  reduction = NULL,
  nCluster = 4
)
}
\arguments{
\item{dataObj}{A \code{Seurat} object.}

\item{platform}{ST or Visium platform}

\item{assay}{assay of the object used}

\item{name}{column name in meta data to store the clustering results}

\item{graphs.name}{naming parameter for stored (S)NN graph}

\item{graphNameSimilarity}{naming parameter for stored similarity}

\item{graphNameDist}{naming parameter for stored distance}

\item{graphNameChain}{naming parameter for stored chain matrix}

\item{nn.name}{Name of knn output on which to run UMAP}

\item{reduction}{dimension reduction matrix in \code{reductions} slot}

\item{nCluster}{the number of clusters}
}
\description{
doBayesSpace
}
\examples{
\dontrun{
  object <- doBayesSpace(
    object,
    platform="ST",
    assay="SCT",
    nCluster=clusterNum)
}

}
