% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSPARK.R
\name{doSpark}
\alias{doSpark}
\title{doSpark}
\usage{
doSpark(
  object,
  n.core = 1,
  genePercentCut = 0.05,
  min_total_counts = 10,
  method = ifelse(length(Cells(object)) < 1000, "spark", "sparkx")
)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{n.core}{number of cores for parallel computing}

\item{genePercentCut}{The percentage of cells that are expressed for analysis}

\item{min_total_counts}{The minimum counts for each cell for filtering}

\item{method}{set automatically based on the sample size: if < 1000 spots,
\code{method} == "spark", otherwise "sparkx".}
}
\value{
a matrix
}
\description{
doSpark
}
