% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSpectrum.R
\name{doSpectrum}
\alias{doSpectrum}
\title{doSpectrum}
\usage{
doSpectrum(
  dataObj,
  reduction.list = NULL,
  graphs = NULL,
  clusterEachModality = TRUE,
  nCluster = 4,
  clusterColumnName = paste0("Spectrum_Cluster", nCluster),
  integrated.name = "SpectrumIntegrated_Similarity",
  nn.name = "SpectrumNeighbor",
  reduction.name = paste0(nn.name, "UMAP"),
  reduction.key = paste0(nn.name, "UMAP_")
)
}
\arguments{
\item{dataObj}{A \code{Seurat} object.}

\item{reduction.list}{list of dimension reduction matrices for integration}

\item{graphs}{graph slot used for similarity calculation if don't define
\code{reduction.list}.}

\item{clusterEachModality}{cluster each modality if set to TRUE. TRUE as
default.}

\item{nCluster}{the number of clusters}

\item{clusterColumnName}{column name to store clustering results}

\item{integrated.name}{integrated name used in graph slot}

\item{nn.name}{Multimodal neighbor object name in neighbors slot}

\item{reduction.name}{Name of projected UMAP to store in the query}

\item{reduction.key}{Value for the projected UMAP key}
}
\description{
doSpectrum
}
