% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DostLearn.R
\name{DostLearn}
\alias{DostLearn}
\title{DostLearn}
\usage{
DostLearn(
  dataObj,
  ipcaObj = NULL,
  imageDimReducName = "ImageFeaturePCA",
  pcaDim_g = 30,
  pcaDim_i = 30,
  geneDimReducName = NULL,
  distanceMethod = c("cosine", "euclidean"),
  spotsCoordinate = NULL,
  platform = ifelse("platform" \%in\% names(dataObj@misc), dataObj@misc$platform,
    "Visium"),
  geneExp = NULL,
  normalizePC = FALSE,
  assay = "SCT",
  dataSlot = c("counts", "data", "scale.data"),
  weights = c("weights_matrix_all", "weights_matrix_pd_md", "physical_distance")
)
}
\arguments{
\item{dataObj}{A \code{Seurat} object}

\item{ipcaObj}{user provided image feature PCA object}

\item{imageDimReducName}{image feature dimension reduction matrix in object}

\item{pcaDim_g}{number of PCs of gene expression dimension reduction matrix}

\item{pcaDim_i}{number of PCs of image feature dimension reduction matrix}

\item{geneDimReducName}{gene expression dimension reduction matrix in object}

\item{distanceMethod}{method for distance calculation, \code{cosine} or
\code{euclidean}}

\item{spotsCoordinate}{spot coordinate information, extract from object if
not provided}

\item{platform}{platform information. either "ST" or "Visium"}

\item{geneExp}{user provided gene expression matrix. i.e. Spark genes}

\item{normalizePC}{logical value to turn on normalization of PC matrix.
FALSE as default}

\item{assay}{assay of gene expression}

\item{dataSlot}{slot of assay of gene expression}

\item{weights}{define of normalization weights, based on 1)
\code{weights_matrix_all} consider gene level correlation in weight;
2) weights from morphological Similarly and physical distance;
3) physical distance only.}
}
\value{
a seurat object
}
\description{
DostLearn
}
\examples{
\dontrun{
object <- DostLearn(
  object,
  pcaDim_i = 20,
  pcaDim_g = 20,
  geneDimReducName = "SCTPCA",
  assay = "SCT",
  platform = "ST",
  distanceMethod = "cosine",
  dataSlot = c("data"),
  weights="weights_matrix_all")
}

}
