% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractImageFeature.R
\name{extract_features_Visium}
\alias{extract_features_Visium}
\title{extract image features for Visium dataset}
\usage{
extract_features_Visium(
  spatialDir,
  patchSize = NULL,
  rawImage = FALSE,
  imgFile = NULL,
  ...
)
}
\arguments{
\item{spatialDir}{the folder contains spatial related files. It's
\code{spatial} if data were generated by Space Ranger.}

\item{patchSize}{the diameter of patches. the spot diameter in
\code{scalefactors_json.json} is used if not defined.}

\item{rawImage}{If or not the raw image is used. FALSE by default.}

\item{imgFile}{the image file used here if \code{rawImage} is TRUE.}

\item{...}{Arguments passed to \code{\link{extract_features}}}
}
\value{
A list of two matrices
}
\description{
extract image features for Visium dataset
}
\examples{
\dontrun{
# Visium platform
imageFeatures.list <-
  extract_features_Visium(
    spatial_path,
    rawImage = T,
    imgFile = rawimg,
    savePatchOnImage = NULL)
imageFeatures <- imageFeatures.list[["ImageFeature"]]
RGBquantile <- imageFeatures.list[["RGBquantile"]]
}
}
