% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DostLearn.R, R/normalizeGeneByImage.R
\name{ExtractGeneExpByNeighbor}
\alias{ExtractGeneExpByNeighbor}
\title{ExtractGeneExpByNeighbor}
\usage{
ExtractGeneExpByNeighbor(geneExp, spotToNeighborList, spotsImageSimilarity)

ExtractGeneExpByNeighbor(geneExp, spotToNeighborList, spotsImageSimilarity)
}
\arguments{
\item{geneExp}{gene expression matrix}

\item{spotToNeighborList}{the output list from \code{\link{FindCoordinateNeighbor}} function}

\item{spotsImageSimilarity}{spot similarity evaluated by image modality}
}
\value{
a gene expression matrix

gene expression matrix weighted by expression matrices of neighbor spots
}
\description{
ExtractGeneExpByNeighbor

Title
}
