% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricFunctions.R
\name{extractNeighborDist}
\alias{extractNeighborDist}
\title{Title}
\usage{
extractNeighborDist(
  dataObj,
  neighborsK = 10,
  sampleN = 20000,
  pcaN = 1:20,
  reduction = "SCTPCA"
)
}
\arguments{
\item{dataObj}{input seurat object}

\item{neighborsK}{number of neighbors to consider}
}
\value{
distance matrix of neighbors
}
\description{
Title
}
