% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNClusters.R
\name{FindClustersForUniroot}
\alias{FindClustersForUniroot}
\title{FindClustersForUniroot}
\usage{
FindClustersForUniroot(x, object, nCluster = 6, graph.name = NULL, ...)
}
\arguments{
\item{x}{numeric value for resolution setting}

\item{object}{A \code{Seurat} object.}

\item{nCluster}{number of clusters}

\item{...}{Arguments passed to \code{\link{FindClusters}}}
}
\description{
FindClustersForUniroot
}
