% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DostLearn.R, R/normalizeGeneByImage.R
\name{FindCoordinateNeighbor}
\alias{FindCoordinateNeighbor}
\title{FindCoordinateNeighbor}
\usage{
FindCoordinateNeighbor(spotsCoordinate, n = 6)

FindCoordinateNeighbor(spotsCoordinate, n = 6)
}
\arguments{
\item{spotsCoordinate}{spot coordinates}

\item{n}{number of neighbor spots}
}
\value{
a list containing neighbors of spots

a list of neighbor spots
}
\description{
FindCoordinateNeighbor

Title
}
