% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNClusters.R
\name{FindNClusters}
\alias{FindNClusters}
\title{Title}
\usage{
FindNClusters(
  object,
  nCluster = 6,
  resolution = 0.8,
  resolutionMin = max(0.2, resolution - 1),
  resolutionMax = min(resolution + 1, 3),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
