% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSNNClusters.R
\name{findSNNClusters}
\alias{findSNNClusters}
\title{Title}
\usage{
findSNNClusters(
  object,
  reduction = "pca",
  graph.name = paste0(reduction, c("_nn", "_snn")),
  nn.name = NULL,
  dims = 1:30,
  resolution = 0.8,
  assay = DefaultAssay(object),
  nCluster = NULL,
  clusterColumnName = if (!is.null(nCluster)) paste0(reduction, "_cluster", nCluster)
    else NULL,
  reduction.name = paste0(reduction, "UMAP"),
  reduction.key = paste0(reduction, "UMAP_"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
