% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMIA.R
\name{getMIA}
\alias{getMIA}
\title{getMIA function of Multimodal intersection analysis (MIA) proposed in
Moncada et al., 2020 (https://doi.org/10.1038/s41587-019-0392-8).}
\usage{
getMIA(singlecellobject, sc_marker = NULL, st_marker = NULL)
}
\arguments{
\item{singlecellobject}{A \code{Seurat} object of scRNAseq data.}

\item{sc_marker}{data frame of marker genes of cell types from scRNAseq data}

\item{st_marker}{data frame of marker genes of spatial clusters from
spatial transcriptomics data}
}
\value{
a dataframe
}
\description{
getMIA function of Multimodal intersection analysis (MIA) proposed in
Moncada et al., 2020 (https://doi.org/10.1038/s41587-019-0392-8).
}
