% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{gg_color_hue}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors}
}
\value{
a color palette
}
\description{
gg_color_hue
}
\examples{
\dontrun{
  gg_color_hue(8)
}
}
