% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricFunctions.R
\name{HighSimilarityNonNeighborPlot}
\alias{HighSimilarityNonNeighborPlot}
\title{Title}
\usage{
HighSimilarityNonNeighborPlot(
  x,
  var.x = "Neighbor.min",
  var.x.ref = "All.Q5",
  var.y = "Neighbor.median",
  var.y.ref = "All.Q5",
  var.x.highlight = 0.5,
  var.y.highlight = 0.5,
  ...
)
}
\arguments{
\item{x}{result table from extractNeighborDist}

\item{...}{}
}
\value{
High Similarity Non-Neighbor Plot
}
\description{
Title
}
