% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityIntegration.R
\name{IntegrationByDistance}
\alias{IntegrationByDistance}
\title{IntegrationByDistance}
\usage{
IntegrationByDistance(
  dataObj = NULL,
  distMatrix = c("SCTBayesSpace_dist_Cluster10", "ImageFeatureBayesSpace_dist_Cluster10"),
  snnMatrix = c("SCTBayesSpace_snn_Cluster10", "ImageFeatureBayesSpace_snn_Cluster10"),
  name = "DistIntegrated",
  graphs.name = paste0(name, c("_nn", "_snn")),
  nn.name = paste0(name, "_Neighbor"),
  integrated.name = paste0(name, "_Dist"),
  integrated.weight.name = paste0(name, "_DistWeight"),
  k.param = 20,
  minDiff = 10^-4
)
}
\arguments{
\item{dataObj}{A \code{Seurat} object.}

\item{distMatrix}{a list of distance Matrices if \code{dataObj} is NULL}

\item{snnMatrix}{a list of SNN Matrices; provided by user or from graphs slot}

\item{name}{naming prefix to store results related to IntegrationByDistance}

\item{graphs.name}{Name of (s)nn output in graphs slot}

\item{nn.name}{Name of knn output on which to run UMAP}

\item{integrated.name}{name of column in meta data to store the clustering
results}

\item{integrated.weight.name}{name of column in meta data to store the
weights of distance}

\item{k.param}{number of k}

\item{minDiff}{minimum resdiue to add when distance calculation}
}
\value{
a seurat object
}
\description{
IntegrationByDistance
}
