% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadImageFeature.R
\name{LoadImageFeature}
\alias{LoadImageFeature}
\title{Load data from raw matrix files}
\usage{
LoadImageFeature(
  countTable,
  imageFeatures,
  positionTable,
  RGBquantile = NULL,
  project = NULL,
  ...
)
}
\arguments{
\item{countTable}{gene (row) by cell (column) matrix of spatial
transcriptomic data generated by any platform}

\item{imageFeatures}{image feature matrix extracted by
\code{extract_features} or provided by user}

\item{positionTable}{matrix containing a table with rows that correspond to
spots, see 'spatial/tissue_positions.csv' in Visium dataset for example}

\item{project}{project name when create seurat object}

\item{...}{Arguments passed to \code{\link{Read10X_h5}}}
}
\value{
A \code{Seurat} object
}
\description{
Load data from raw matrix files
}
\examples{
\dontrun{
countTable <- read.table(GeneByCellFile)
imageFeatures <- read.csv(ImageFeatureFile)
positionTable <- read.table(PositionTableFile)
LoadImageFeature(countTable = countTable, imageFeatures = imageFeatures,
positionTable = positionTable, project = "test")
}

}
\concept{preprocessing}
