% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadImageFeature.R
\name{LoadImageFeatureVisium}
\alias{LoadImageFeatureVisium}
\title{Load Visium data}
\usage{
LoadImageFeatureVisium(
  dataDir,
  filename = "filtered_feature_bc_matrix.h5",
  imageFeatures,
  assay = "Spatial",
  slice = "slice1",
  filter.matrix = TRUE,
  to.upper = FALSE,
  image = NULL,
  ...
)
}
\arguments{
\item{dataDir}{a path/folder which contains the dataset generated by
SpaceRanger}

\item{filename}{Directory containing the H5 file specified by \code{filename}
and the image data in a subdirectory called \code{spatial}}

\item{imageFeatures}{image feature matrix extracted from ImageFeature.R or
provided by user}

\item{assay}{assay name for storing gene expression matrix in seurat object,
default value: "Spatial"}

\item{slice}{Name for the stored image of the tissue slice, default value:
"slice1"}

\item{filter.matrix}{Only keep spots that have been determined to be over
tissue, default value: TRUE}

\item{to.upper}{Converts all feature names to upper case. This can provide an
approximate conversion of mouse to human gene names which can be useful in an
explorative analysis. For cross-species comparisons, orthologous genes should
be identified across species and used instead.}

\item{image}{An object of class VisiumV1. Typically, an output from \code{\link{Read10X_Image}}}

\item{...}{Arguments passed to \code{\link{Read10X_h5}}}
}
\value{
A \code{Seurat} object
}
\description{
Load Visium data
}
\examples{
\dontrun{
data_dir <- 'path/to/data/directory'
list.files(data_dir) # Should show filtered_feature_bc_matrix.h5
imageFeatures <- read.csv(ImageFeatureFile)
LoadImageFeatureVisium(data.dir = data_dir, filename = h5_file,
imageFeatures = imageFeatures)
}

}
\concept{preprocessing}
