% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricFunctions.R
\name{ModalityConcordancePlot}
\alias{ModalityConcordancePlot}
\title{Title}
\usage{
ModalityConcordancePlot(
  dataObj,
  i = 1:20,
  sampleN = 20000,
  normalize = TRUE,
  diffCutoff = 5,
  geneReduction = "SCTPCA",
  ImageReduction = "ImageFeaturePCA",
  ...
)
}
\arguments{
\item{dataObj}{seurat object}

\item{i}{number of PCs to use}

\item{sampleN}{number of samples in resampling}

\item{normalize}{normalize distance by min distance, True or False}

\item{diffCutoff}{cutoff value to highlight spot pairs}

\item{...}{}
}
\value{
Modality Concordance Plot
}
\description{
Title
}
