% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultimodalImageWorkflow.R
\name{MultimodalImageWorkflowVisium}
\alias{MultimodalImageWorkflowVisium}
\title{integrating and processing for Visium dataset}
\usage{
MultimodalImageWorkflowVisium(
  object,
  normalizeMethod = "SCT",
  ImageFeaturePCs = NULL,
  pcaDim_s = 30,
  pcaDim_i = 30,
  geneResolution = 0.8,
  imageFeatureResolution = geneResolution,
  imageAndGeneResolution = geneResolution,
  cia.nf = 20,
  IntNMF.nf = 20,
  genePercentCut = 0.1,
  imagePercentCut = 0.3,
  MultimodalMethod = c("WNN", "MCIA", "IntNMF", "tICA"),
  WnnImageWeightFactor = 1
)
}
\arguments{
\item{WnnImageWeightFactor}{}
}
\value{

}
\description{
integrating and processing for Visium dataset
}
