% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultimodalIntegrationVisium.R
\name{MultiModalIntegrationVisium}
\alias{MultiModalIntegrationVisium}
\title{integrating and processing for Visium dataset}
\usage{
MultiModalIntegrationVisium(
  object,
  normalizeMethod = c("SCT", "log"),
  reduction.list = list("SCTPCA", "ImageFeaturePCA"),
  pcaDim_s = 30,
  pcaDim_i = 30,
  resolution = 0.8,
  nCluster = NULL,
  clusterColumnName = if (!is.null(nCluster)) paste0("wnn_cluster", nCluster) else NULL,
  cia.nf = 20,
  IntNMF.nf = 20,
  MultimodalMethod = c("WNN", "MCIA", "IntNMF", "tICA"),
  WnnImageWeightFactor = 1,
  ...
)
}
\arguments{
\item{normalizeMethod}{}
}
\value{

}
\description{
integrating and processing for Visium dataset
}
