% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultimodalPreProcess.R
\name{MultimodalPreProcess}
\alias{MultimodalPreProcess}
\title{Title}
\usage{
MultimodalPreProcess(
  object,
  normalizeMethod = c("SCT", "log"),
  pcaDim_s = 30,
  pcaDim_i = 30,
  pcaDim_c = 6,
  DimReducMethod = c("PCA", "SpatialPCA"),
  genePercentCut = 0.05,
  imagePercentCut = 0.05,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
