% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityIntegration.R
\name{NeighborDist}
\alias{NeighborDist}
\title{NeighborDist}
\usage{
NeighborDist(distMatrix, neighborList, removeFirst = TRUE)
}
\arguments{
\item{distMatrix}{distance matrix}

\item{neighborList}{list of neighbor}

\item{removeFirst}{logical value to remove the spot itself. TRUE as default}
}
\value{
a dist dataframe
}
\description{
NeighborDist
}
