% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricFunctions.R
\name{plotDistConsistency}
\alias{plotDistConsistency}
\title{Title}
\usage{
plotDistConsistency(dataObj, i = 1:20, sampleN = 5000, normalize = TRUE, ...)
}
\arguments{
\item{i}{number of PCs to use}

\item{sampleN}{number of samples in resampling}

\item{normalize}{normalize distance by min distance, True or False}

\item{...}{}
}
\description{
Title
}
