% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReProcessAfterRenormByImage.R
\name{ReProcessAfterRenormByImage}
\alias{ReProcessAfterRenormByImage}
\title{Title}
\usage{
ReProcessAfterRenormByImage(
  object,
  normalizeMethod = c("SCT", "log"),
  renormbyimage = c("GeneExp", "GeneExpPCs"),
  pcaDim_s = 30,
  pcaDim_i = 30,
  ipcaObj = NULL,
  distanceMethod = "cosine",
  spotsCoordinate = NULL,
  coordinateNeighborN = 6,
  geneExp = NULL,
  DimReducMethod = c("PCA", "SpatialPCA"),
  genePercentCut = 0.1,
  imagePercentCut = 0.3,
  geneResolution = 0.8,
  imageFeatureResolution = geneResolution,
  sparkversion = "spark",
  numCores_spark = 1,
  gene.number = 3000,
  customGenelist = NULL,
  min.loctions = 30,
  min.features = 30
)
}
\arguments{
\item{min.features}{}
}
\value{

}
\description{
Title
}
