% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDimReduc.R
\name{RunDimReduc}
\alias{RunDimReduc}
\title{RunDimReduc}
\usage{
RunDimReduc(
  object,
  DimReducMethod = "PCA",
  assay = "SCT",
  percentCut = 0.05,
  pcaDim = 30,
  customGenes = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{DimReducMethod}{dimension reduction method. Can be either \code{PCA}
or \code{SpatialPCA}}

\item{assay}{assay used for dimension reduction analysis. i.e., \code{SCT}
for gene matrix when preprocessed by \code{SCTransform}}

\item{percentCut}{cutoff value of percentage of spots of their
values larger than minimum value}

\item{pcaDim}{number of PCs when running dimension reduction}

\item{customGenes}{vector of custom genes to run dimension reduction step}

\item{...}{potentially needed Arguments}
}
\value{
a seurat object
}
\description{
RunDimReduc
}
