% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricFunctions.R
\name{showSpotsDiff}
\alias{showSpotsDiff}
\title{Title}
\usage{
showSpotsDiff(
  x,
  varName1 = "Neighbor.mean",
  varName2 = "Random.mean",
  main = paste0("Difference between ", varName1, " and ", varName2),
  ...
)
}
\arguments{
\item{x}{result table from extractNeighborDist}

\item{varName1}{var for Neighbor distance. default is Neighbor.mean}

\item{varName2}{var for distance to compare with. default is Random.mean}

\item{...}{}
}
\description{
Title
}
