% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialAwareProcess.R
\name{SpatialAwareProcess}
\alias{SpatialAwareProcess}
\title{SpatialAwareProcess}
\usage{
SpatialAwareProcess(
  object,
  platform = c("ST", "Visium"),
  assay = c("Spatial", "SCT", "ImageFeature"),
  slot = c("counts", "data", "scale.data"),
  SpatialAwareMethod = c("SpatialPCA", "BayesSpace", "stLearn"),
  clusterNum = NULL,
  pcaDim_g = 30,
  pcaDim_i = 30,
  percentCut = 0.2,
  customGenes = NULL,
  distanceMethod = c("cosine", "euclidean"),
  stLearnweights = c("weights_matrix_all", "weights_matrix_pd_md", "physical_distance")
)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{platform}{platform of ST data}

\item{assay}{Assay to use in construction of (S)NN}

\item{slot}{slot of assays used for processing}

\item{SpatialAwareMethod}{spatial-aware processing method, one of
"BayesSpace", "SpatialPCA", and "stLearn"}

\item{clusterNum}{number of clusters}

\item{pcaDim_g}{number of PCs when running dimension reduction on gene
expression (spatial) matrix, 30 as default}

\item{pcaDim_i}{number of PCs when running dimension reduction on image
feature matrix, 30 as default}

\item{percentCut}{cutoff value of percentage of spots of their
values larger than minimum value for filtering genes after normalization.}

\item{customGenes}{custom gene sets when performing SpatialPCA, default NULL.}

\item{distanceMethod}{method for distance calculation, \code{cosine} or
\code{euclidean}}

\item{stLearnweights}{define of normalization weights, based on 1)
\code{weights_matrix_all} consider gene level correlation in weight;
2) weights from morphological Similarly and physical distance;
3) physical distance only.}
}
\value{
A \code{Seurat} object
}
\description{
SpatialAwareProcess
}
\examples{
\dontrun{
object <-
  SpatialAwareProcess(
    object,
    patform = "ST",
    assay = "SCT",
    slot = "data",
    SpatialAwareMethod = "stLearn",
    clusterNum = 8,
    pcaDim_g = 20,
    pcaDim_i = 20,
    percentCut = 0.2,
    customGenes = NULL,
    distanceMethod = "cosine",
    stLearnweights = "weights_matrix_all"
 )
}

}
