% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCAWorkflow.R
\name{SpatialPCAWorkflow}
\alias{SpatialPCAWorkflow}
\title{Title}
\usage{
SpatialPCAWorkflow(
  SpatialPCAObj,
  SpatialPCnum = 30,
  kerneltype = "gaussian",
  bandwidthtype = "SJ",
  fast = FALSE
)
}
\arguments{
\item{fast}{}
}
\value{

}
\description{
Title
}
