% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{st_voronoi_point}
\alias{st_voronoi_point}
\title{st_voronoi_point}
\usage{
st_voronoi_point(points)
}
\arguments{
\item{points}{sf data frame of points}
}
\value{
a sf data frame
}
\description{
st_voronoi_point
}
\examples{
\dontrun{
  location <- GetTissueCoordinates(object)
  p <- st_as_sf(data.frame(x=location$y, y=dplyr::desc(location$x),
  A=1:nrow(location)), coords=1:2)
  v <- st_voronoi_point(p)
}
}
