% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{AllSpatialDimPlot}
\alias{AllSpatialDimPlot}
\title{AllSpatialDimPlot}
\usage{
AllSpatialDimPlot(
  object,
  cluster.frame = TRUE,
  cluster.highlight = NULL,
  col = NULL,
  ...
)
}
\description{
AllSpatialDimPlot
}
