% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBayesSpace.R
\name{bayesSpaceChainMatrixSimilarity}
\alias{bayesSpaceChainMatrixSimilarity}
\title{bayesSpaceChainMatrixSimilarity}
\usage{
bayesSpaceChainMatrixSimilarity(bayesSpaceResultChainMatrix)
}
\arguments{
\item{bayesSpaceResultChainMatrix}{MCMC chain saved in object}
}
\description{
bayesSpaceChainMatrixSimilarity
}
