% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityIntegration.R
\name{distToNeighborObj}
\alias{distToNeighborObj}
\title{distToNeighborObj}
\usage{
distToNeighborObj(distMatrix, k.param = 20)
}
\arguments{
\item{distMatrix}{distance matrix}

\item{k.param}{number of k}
}
\description{
distToNeighborObj
}
