% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageFeature.R
\name{extract_features}
\alias{extract_features}
\title{extract image features}
\usage{
extract_features(
  imgFile,
  positionTable,
  patchSize = NULL,
  scaleFactor = 1,
  rawImage = FALSE,
  savePatchOnImage = NULL
)
}
\arguments{
\item{savePatchOnImage}{}
}
\value{

}
\description{
extract image features
}
