% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageFeature.R
\name{extract_features_Visium}
\alias{extract_features_Visium}
\title{extract image features for Visium dataset}
\usage{
extract_features_Visium(
  spatialDir,
  patchSize = NULL,
  rawImage = FALSE,
  imgFile = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
extract image features for Visium dataset
}
