% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dostLearn.R
\name{ExtractGeneExpByNeighbor}
\alias{ExtractGeneExpByNeighbor}
\title{ExtractGeneExpByNeighbor}
\usage{
ExtractGeneExpByNeighbor(geneExp, spotToNeighborList, spotsImageSimilarity)
}
\arguments{
\item{geneExp}{gene expression matrix}

\item{spotToNeighborList}{neighbor spot list returned by
\code{FindCoordinateNeighbor}}

\item{spotsImageSimilarity}{similarity matrix between spots calculated with
the neighbor image features}
}
\value{
a gene expression matrix
}
\description{
ExtractGeneExpByNeighbor
}
