% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeGeneByImage.R
\name{ExtractGeneExpByNeighbor}
\alias{ExtractGeneExpByNeighbor}
\title{Title}
\usage{
ExtractGeneExpByNeighbor(geneExp, spotToNeighborList, spotsImageSimilarity)
}
\arguments{
\item{spotsImageSimilarity}{}
}
\value{

}
\description{
Title
}
