% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricFunctions.R
\name{extractNeighborDist}
\alias{extractNeighborDist}
\title{Title}
\usage{
extractNeighborDist(dataObj, neighborsK = 10)
}
\arguments{
\item{neighborsK}{number of neighbors to consider}
}
\description{
Title
}
