% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DostLearn.R
\name{FindCoordinateNeighbor}
\alias{FindCoordinateNeighbor}
\title{FindCoordinateNeighbor}
\usage{
FindCoordinateNeighbor(spotsCoordinate, n = 6)
}
\arguments{
\item{spotsCoordinate}{coordinate information of spots}

\item{n}{number of neighbor spots}
}
\value{
a list containing neighbors of spots
}
\description{
FindCoordinateNeighbor
}
