% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dostLearn.R
\name{FindCoordinateNeighbor}
\alias{FindCoordinateNeighbor}
\title{FindCoordinateNeighbor}
\usage{
FindCoordinateNeighbor(spotsCoordinate, n = 6)
}
\arguments{
\item{spotsCoordinate}{coordinate information of spots}

\item{n}{number of neighbor spots}
}
\description{
FindCoordinateNeighbor
}
