% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSNNClusters.R
\name{findSNNClusters}
\alias{findSNNClusters}
\title{findSNNClusters}
\usage{
findSNNClusters(
  object,
  reduction = "pca",
  graph.name = paste0(reduction, c("_nn", "_snn")),
  nn.name = NULL,
  dims = 1:30,
  resolution = 0.8,
  assay = DefaultAssay(object),
  nCluster = NULL,
  k.param = 20,
  clusterColumnName = if (!is.null(nCluster)) paste0(reduction, "_cluster", nCluster)
    else NULL,
  reduction.name = ifelse(!is.null(reduction), paste0(reduction, "UMAP"), paste0(nn.name,
    "UMAP")),
  reduction.key = ifelse(!is.null(reduction), paste0(reduction, "UMAP_"), paste0(nn.name,
    "UMAP_")),
  RunUMAP = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{Seurat} object.}

\item{reduction}{dimension reduction method}

\item{graph.name}{naming parameter for stored (S)NN graph}

\item{nn.name}{Name of knn output on which to run UMAP}

\item{dims}{which dimensions to use as input features}

\item{resolution}{value of the resolution parameter, use a value above
(below) 1.0 if you want to obtain a larger (smaller) number of communities.}

\item{assay}{Assay to use in construction of (S)NN; used only when dims is
NULL}

\item{nCluster}{number of clusters}

\item{k.param}{Defines k for the k-nearest neighbor algorithm}

\item{clusterColumnName}{column name in meta data to store the clustering
results}

\item{reduction.name}{name of projected UMAP to store in the query}

\item{reduction.key}{value for the projected UMAP key}

\item{RunUMAP}{logical value to define if or not to run RunUMAP}

\item{...}{Arguments passed to \code{\link{FindNClusters}}}
}
\value{
a seurat object
}
\description{
findSNNClusters
}
\examples{
\dontrun{
object <- findSNNClusters(
  object,
  dims = 1:20,
  reduction = "SCTPCA",
  nCluster = clusterNum,
  resolutionMax = 3
  )
}

}
