% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultimodalIntegrationVisium.R
\name{MultiModalIntegrationVisium}
\alias{MultiModalIntegrationVisium}
\title{integrating and processing for Visium dataset}
\usage{
MultiModalIntegrationVisium(
  object,
  normalizeMethod = c("SCT", "log"),
  DimReducMethod = c("PCA", "SpatialPCA"),
  pcaDim_s = 30,
  pcaDim_i = 30,
  imageAndGeneResolution = 0.8,
  cia.nf = 20,
  IntNMF.nf = 20,
  MultimodalMethod = c("WNN", "MCIA", "IntNMF", "tICA"),
  WnnImageWeightFactor = 1
)
}
\arguments{
\item{DimReducMethod}{}
}
\value{

}
\description{
integrating and processing for Visium dataset
}
