% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultimodalPreProcess.R
\name{MultimodalPreProcess}
\alias{MultimodalPreProcess}
\title{Title}
\usage{
MultimodalPreProcess(
  object,
  GeneFeaturePCobj = NULL,
  ImageFeaturePCobj = NULL,
  normalizeMethod = c("SCT", "log"),
  pcaDim_s = 30,
  pcaDim_i = 30,
  genePercentCut = 0.1,
  imagePercentCut = 0.3,
  geneResolution = 0.8,
  imageFeatureResolution = geneResolution
)
}
\arguments{
\item{imageFeatureResolution}{}
}
\value{

}
\description{
Title
}
