% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeGeneByImage.R
\name{NormalizeGeneByImage}
\alias{NormalizeGeneByImage}
\title{Title}
\usage{
NormalizeGeneByImage(
  dataObj,
  ipcaObj = NULL,
  nPCs = 30,
  distanceMethod = "cosine",
  spotsCoordinate = NULL,
  coordinateNeighborN = 4,
  geneExp = NULL
)
}
\arguments{
\item{geneExp}{}
}
\value{

}
\description{
Title
}
