% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeGeneByImage.R
\name{NormalizeGeneByImage}
\alias{NormalizeGeneByImage}
\title{Title}
\usage{
NormalizeGeneByImage(
  dataObj,
  ipcaObj = NULL,
  imageDimReducName = "ImageFeaturePCA",
  pcaDim_i = 30,
  geneDimReducName = NULL,
  pcaDim_g = 30,
  distanceMethod = c("cosine", "euclidean"),
  spotsCoordinate = NULL,
  platform = ifelse("platform" \%in\% names(dataObj@misc), dataObj@misc$platform,
    "Visium"),
  geneExp = NULL,
  normalizePC = FALSE,
  assay = "SCT",
  dataSlot = c("counts", "data", "scale.data"),
  weights = c("weights_matrix_all", "weights_matrix_pd_md", "physical_distance")
)
}
\arguments{
\item{geneExp}{}
}
\description{
Title
}
