% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeGeneByImage.R
\name{NormalizeGeneByImage}
\alias{NormalizeGeneByImage}
\title{Title}
\usage{
NormalizeGeneByImage(
  dataObj,
  ipcaObj = NULL,
  imageDimReducName = "ImageFeaturePCA",
  pcaDim_i = 30,
  distanceMethod = "cosine",
  spotsCoordinate = NULL,
  coordinateNeighborN = 4,
  geneExp = NULL,
  geneDimReducName = NULL,
  assay = "SCT",
  dataSlot = c("data", "scale.data")
)
}
\arguments{
\item{geneExp}{}
}
\value{

}
\description{
Title
}
