% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{PreProcessing}
\alias{PreProcessing}
\title{MultiModalPreProcessing}
\usage{
PreProcessing(
  object,
  normalizeMethod = c("SCT", "log"),
  pcaDim_s = 30,
  pcaDim_i = 30,
  pcaDim_c = 6,
  DimReducMethod = "PCA",
  prefiltergenePercentCut = 0.05,
  prefilterimagePercentCut = 0.05,
  genePercentCut = 0.05,
  imagePercentCut = 0.05,
  customGenes = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{Seurat} object prepared by \code{LoadImageFeature} or
\code{LoadImageFeatureVisium}}

\item{normalizeMethod}{the normalization method for gene expression matrix.
\code{log} for \code{LogNormalize}, \code{SCT} for \code{SCTransform}.}

\item{pcaDim_s}{number of PCs when running dimension reduction on gene
expression (spatial) matrix, 30 as default}

\item{pcaDim_i}{number of PCs when running dimension reduction on image
feature matrix, 30 as default}

\item{pcaDim_c}{number of PCs when running dimension reduction on RGB
quantile matrix, 6 as default}

\item{DimReducMethod}{dimension reduction method. Can be either \code{PCA}
or \code{SpatialPCA}}

\item{prefiltergenePercentCut}{cutoff value of percentage of spots of their
values larger than minimum value for prefiltering genes.}

\item{prefilterimagePercentCut}{cutoff value of percentage of spots of their
values larger than minimum value for prefiltering image features.}

\item{genePercentCut}{cutoff value of percentage of spots of their
values larger than minimum value for filtering genes after normalization.}

\item{imagePercentCut}{cutoff value of percentage of spots of their
values larger than minimum value for prefiltering image features after
normalization.}

\item{customGenes}{custom gene list provide by user. If is not \code{NULL},
this list will be used instead of the most variable genes.}

\item{...}{Arguments passed to \code{\link{RunDimReduc}}}
}
\value{
A \code{Seurat} object
}
\description{
MultiModalPreProcessing
}
\examples{
\dontrun{
object <-
  PreProcessing(
    object,
    pcaDim_s = 20,
    pcaDim_i = 20,
    pcaDim_c = 6,
    normalizeMethod = "SCT",
    DimReducMethod = "PCA",
    prefiltergenePercentCut=0,
    prefilterimagePercentCut=0,
    genePercentCut = 0.25,
    imagePercentCut = 0.2
 )
}

}
