% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDimReduc.R
\name{runDimReduc}
\alias{runDimReduc}
\title{Title}
\usage{
runDimReduc(
  object,
  DimReducMethod = "PCA",
  assay = "SCT",
  percentCut = 0.05,
  pcaDim = 30,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
