% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPCAWorkflow.R
\name{SpatialPCAWorkflow}
\alias{SpatialPCAWorkflow}
\title{SpatialPCAWorkflow}
\usage{
SpatialPCAWorkflow(
  SpatialPCAObj,
  SpatialPCnum = 30,
  kerneltype = "gaussian",
  bandwidthtype = "SJ",
  fast = FALSE
)
}
\arguments{
\item{SpatialPCAObj}{A \code{SpatialPCA} object}

\item{SpatialPCnum}{number of PCs when running SpatialPCA}

\item{kerneltype}{The type of kernel to be used, either "gaussian" for
gaussian kernel, or "cauchy" for cauchy kernel, or "quadratic" for rational
quadratic kernel.}

\item{bandwidthtype}{The type of bandwidth to be used in Gaussian kernel,
"SJ" for Sheather & Jones (1991) method (usually used in small sample size
datasets), "Silverman" for Silverman's ‘rule of thumb’ method (1986)(usually
used in large sample size datasets).}

\item{fast}{Select "TRUE" to accrelerate the algorithm by performing
low-rank approximation on the kernel matrix, otherwise "FALSE" for
calculation without low-rank approximation on the kernel matrix.}
}
\description{
SpatialPCAWorkflow
}
\examples{
\dontrun{SPCAobj <- SpatialPCAWorkflow(SPCAobj, SpatialPCnum = pcaDim)}

}
